+++ packages/go_router/test/test_helpers.dart
@@ -127,10 +127,12 @@ class GoRouterPopSpy extends GoRouter {
   GoRouterPopSpy({required super.routes});
 
   bool popped = false;
+  Object? poppedResult;
 
   @override
   void pop<T extends Object?>([T? result]) {
     popped = true;
+    poppedResult = result;
   }
 }