+++ packages/pigeon/platform_tests/alternate_language_test_plugin/android/src/main/java/com/example/alternate_language_test_plugin/AlternateLanguageTestPlugin.java
@@ -13,6 +13,8 @@ import com.example.alternate_language_test_plugin.CoreTests.FlutterIntegrationCo
 import com.example.alternate_language_test_plugin.CoreTests.HostIntegrationCoreApi;
 import com.example.alternate_language_test_plugin.CoreTests.Result;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import java.util.List;
+import java.util.Map;
 
 /** This plugin handles the native side of the integration tests in example/integration_test/. */
 public class AlternateLanguageTestPlugin implements FlutterPlugin, HostIntegrationCoreApi {
@@ -155,11 +157,185 @@ public class AlternateLanguageTestPlugin implements FlutterPlugin, HostIntegrati
   }
 
   @Override
+  public void callFlutterEchoAllTypes(@NonNull AllTypes everything, Result<AllTypes> result) {
+    flutterApi.echoAllTypes(
+        everything,
+        new FlutterIntegrationCoreApi.Reply<AllTypes>() {
+          public void reply(AllTypes value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterSendMultipleNullableTypes(
+      @Nullable Boolean aNullableBool,
+      @Nullable Long aNullableInt,
+      @Nullable String aNullableString,
+      Result<AllNullableTypes> result) {
+    flutterApi.sendMultipleNullableTypes(
+        aNullableBool,
+        aNullableInt,
+        aNullableString,
+        new FlutterIntegrationCoreApi.Reply<AllNullableTypes>() {
+          public void reply(AllNullableTypes value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoBool(@NonNull Boolean aBool, Result<Boolean> result) {
+    flutterApi.echoBool(
+        aBool,
+        new FlutterIntegrationCoreApi.Reply<Boolean>() {
+          public void reply(Boolean value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoInt(@NonNull Long anInt, Result<Long> result) {
+    flutterApi.echoInt(
+        anInt,
+        new FlutterIntegrationCoreApi.Reply<Long>() {
+          public void reply(Long value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoDouble(@NonNull Double aDouble, Result<Double> result) {
+    flutterApi.echoDouble(
+        aDouble,
+        new FlutterIntegrationCoreApi.Reply<Double>() {
+          public void reply(Double value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
   public void callFlutterEchoString(@NonNull String aString, Result<String> result) {
     flutterApi.echoString(
         aString,
         new FlutterIntegrationCoreApi.Reply<String>() {
           public void reply(String value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoUint8List(@NonNull byte[] aList, Result<byte[]> result) {
+    flutterApi.echoUint8List(
+        aList,
+        new FlutterIntegrationCoreApi.Reply<byte[]>() {
+          public void reply(byte[] value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoList(@NonNull List<Object> aList, Result<List<Object>> result) {
+    flutterApi.echoList(
+        aList,
+        new FlutterIntegrationCoreApi.Reply<List<Object>>() {
+          public void reply(List<Object> value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoMap(
+      @NonNull Map<String, Object> aMap, Result<Map<String, Object>> result) {
+    flutterApi.echoMap(
+        aMap,
+        new FlutterIntegrationCoreApi.Reply<Map<String, Object>>() {
+          public void reply(Map<String, Object> value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoNullableBool(@Nullable Boolean aBool, Result<Boolean> result) {
+    flutterApi.echoNullableBool(
+        aBool,
+        new FlutterIntegrationCoreApi.Reply<Boolean>() {
+          public void reply(Boolean value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoNullableInt(@Nullable Long anInt, Result<Long> result) {
+    flutterApi.echoNullableInt(
+        anInt,
+        new FlutterIntegrationCoreApi.Reply<Long>() {
+          public void reply(Long value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoNullableDouble(@Nullable Double aDouble, Result<Double> result) {
+    flutterApi.echoNullableDouble(
+        aDouble,
+        new FlutterIntegrationCoreApi.Reply<Double>() {
+          public void reply(Double value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoNullableString(@Nullable String aString, Result<String> result) {
+    flutterApi.echoNullableString(
+        aString,
+        new FlutterIntegrationCoreApi.Reply<String>() {
+          public void reply(String value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoNullableUint8List(@Nullable byte[] aList, Result<byte[]> result) {
+    flutterApi.echoNullableUint8List(
+        aList,
+        new FlutterIntegrationCoreApi.Reply<byte[]>() {
+          public void reply(byte[] value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoNullableList(
+      @Nullable List<Object> aList, Result<List<Object>> result) {
+    flutterApi.echoNullableList(
+        aList,
+        new FlutterIntegrationCoreApi.Reply<List<Object>>() {
+          public void reply(List<Object> value) {
+            result.success(value);
+          }
+        });
+  }
+
+  @Override
+  public void callFlutterEchoNullableMap(
+      @Nullable Map<String, Object> aMap, Result<Map<String, Object>> result) {
+    flutterApi.echoNullableMap(
+        aMap,
+        new FlutterIntegrationCoreApi.Reply<Map<String, Object>>() {
+          public void reply(Map<String, Object> value) {
             result.success(value);
           }
         });