+++ packages/pigeon/platform_tests/test_plugin/ios/Classes/TestPlugin.swift
@@ -24,79 +24,79 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
   // MARK: HostIntegrationCoreApi implementation
 
   func noop() {
+
   }
 
-  func echoAllTypes(everything: AllTypes) -> AllTypes {
+  func echo(_ everything: AllTypes) -> AllTypes {
     return everything
   }
 
-  func echoAllNullableTypes(everything: AllNullableTypes?) -> AllNullableTypes? {
+  func echo(_ everything: AllNullableTypes?) -> AllNullableTypes? {
     return everything
   }
 
-  func throwError() {
-    // TODO(stuartmorgan): Implement this. See
-    // https://github.com/flutter/flutter/issues/112483
+  func throwError() throws {
+    throw ErrType.thrownErrow
   }
 
-  func echoInt(anInt: Int32) -> Int32 {
+  func echo(_ anInt: Int32) -> Int32 {
     return anInt
   }
 
-  func echoDouble(aDouble: Double) -> Double {
+  func echo(_ aDouble: Double) -> Double {
     return aDouble
   }
 
-  func echoBool(aBool: Bool) -> Bool {
+  func echo(_ aBool: Bool) -> Bool {
     return aBool
   }
 
-  func echoString(aString: String) -> String {
+  func echo(_ aString: String) -> String {
     return aString
   }
 
-  func echoUint8List(aUint8List: FlutterStandardTypedData) -> FlutterStandardTypedData {
+  func echo(_ aUint8List: FlutterStandardTypedData) -> FlutterStandardTypedData {
     return aUint8List
   }
 
-  func echoObject(anObject: Any) -> Any {
+  func echo(_ anObject: Any) -> Any {
     return anObject
   }
 
-  func extractNestedNullableString(wrapper: AllNullableTypesWrapper) -> String? {
+  func extractNestedNullableString(from wrapper: AllNullableTypesWrapper) -> String? {
     return wrapper.values.aNullableString;
   }
 
-  func createNestedNullableString(nullableString: String?) -> AllNullableTypesWrapper {
+  func createNestedObject(with nullableString: String?) -> AllNullableTypesWrapper {
     return AllNullableTypesWrapper(values: AllNullableTypes(aNullableString: nullableString))
   }
 
-  func sendMultipleNullableTypes(aNullableBool: Bool?, aNullableInt: Int32?, aNullableString: String?) -> AllNullableTypes {
+  func sendMultipleNullableTypes(aBool aNullableBool: Bool?, anInt aNullableInt: Int32?, aString aNullableString: String?) -> AllNullableTypes {
     let someThings = AllNullableTypes(aNullableBool: aNullableBool, aNullableInt: aNullableInt, aNullableString: aNullableString)
     return someThings
   }
 
-  func echoNullableInt(aNullableInt: Int32?) -> Int32? {
+  func echo(_ aNullableInt: Int32?) -> Int32? {
     return aNullableInt
   }
 
-  func echoNullableDouble(aNullableDouble: Double?) -> Double? {
+  func echo(_ aNullableDouble: Double?) -> Double? {
     return aNullableDouble
   }
 
-  func echoNullableBool(aNullableBool: Bool?) -> Bool? {
+  func echo(_ aNullableBool: Bool?) -> Bool? {
     return aNullableBool
   }
 
-  func echoNullableString(aNullableString: String?) -> String? {
+  func echo(_ aNullableString: String?) -> String? {
     return aNullableString
   }
 
-  func echoNullableUint8List(aNullableUint8List: FlutterStandardTypedData?) -> FlutterStandardTypedData? {
+  func echo(_ aNullableUint8List: FlutterStandardTypedData?) -> FlutterStandardTypedData? {
     return aNullableUint8List
   }
 
-  func echoNullableObject(aNullableObject: Any?) -> Any? {
+  func echo(_ aNullableObject: Any?) -> Any? {
     return aNullableObject
   }
 
@@ -104,7 +104,7 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     completion()
   }
 
-  func echoAsyncString(aString: String, completion: @escaping (String) -> Void) {
+  func echoAsync(_ aString: String, completion: @escaping (String) -> Void) {
     completion(aString)
   }
 
@@ -114,9 +114,82 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     }
   }
 
-  func callFlutterEchoString(aString: String, completion: @escaping (String) -> Void) {
-    flutterAPI.echoString(aString: aString) { flutterString in
-      completion(flutterString)
+  func callFlutterEcho(_ everything: AllTypes, completion: @escaping (AllTypes) -> Void) {
+      flutterAPI.echo(everything) { completion($0) }
+  }
+
+  func callFlutterSendMultipleNullableTypes(
+    aBool aNullableBool: Bool?,
+    anInt aNullableInt: Int32?,
+    aString aNullableString: String?,
+    completion: @escaping (AllNullableTypes) -> Void
+  ) {
+    flutterAPI.sendMultipleNullableTypes(
+      aBool: aNullableBool,
+      anInt: aNullableInt,
+      aString: aNullableString
+    ) {
+      completion($0)
     }
   }
+
+  func callFlutterEcho(_ aBool: Bool, completion: @escaping (Bool) -> Void) {
+    flutterAPI.echo(aBool) { completion($0) }
+  }
+
+  func callFlutterEcho(_ anInt: Int32, completion: @escaping (Int32) -> Void) {
+    flutterAPI.echo(anInt) { completion($0) }
+  }
+
+  func callFlutterEcho(_ aDouble: Double, completion: @escaping (Double) -> Void) {
+    flutterAPI.echo(aDouble) { completion($0) }
+  }
+
+  func callFlutterEcho(_ aString: String, completion: @escaping (String) -> Void) {
+    flutterAPI.echo(aString) { completion($0) }
+  }
+
+  func callFlutterEcho(_ aList: FlutterStandardTypedData, completion: @escaping (FlutterStandardTypedData) -> Void) {
+    flutterAPI.echo(aList) { completion($0) }
+  }
+
+  func callFlutterEcho(_ aList: [Any?], completion: @escaping ([Any?]) -> Void) {
+    flutterAPI.echo(aList) { completion($0) }
+  }
+
+  func callFlutterEcho(_ aMap: [String? : Any?], completion: @escaping ([String? : Any?]) -> Void) {
+    flutterAPI.echo(aMap) { completion($0) }
+  }
+
+  func callFlutterEchoNullable(_ aBool: Bool?, completion: @escaping (Bool?) -> Void) {
+    flutterAPI.echoNullable(aBool) { completion($0) }
+  }
+
+  func callFlutterEchoNullable(_ anInt: Int32?, completion: @escaping (Int32?) -> Void) {
+    flutterAPI.echoNullable(anInt) { completion($0) }
+  }
+
+  func callFlutterEchoNullable(_ aDouble: Double?, completion: @escaping (Double?) -> Void) {
+    flutterAPI.echoNullable(aDouble) { completion($0) }
+  }
+
+  func callFlutterEchoNullable(_ aString: String?, completion: @escaping (String?) -> Void) {
+    flutterAPI.echoNullable(aString) { completion($0) }
+  }
+
+  func callFlutterEchoNullable(_ aList: FlutterStandardTypedData?, completion: @escaping (FlutterStandardTypedData?) -> Void) {
+    flutterAPI.echoNullable(aList) { completion($0) }
+  }
+
+  func callFlutterEchoNullable(_ aList: [Any?]?, completion: @escaping ([Any?]?) -> Void) {
+    flutterAPI.echoNullable(aList) { completion($0) }
+  }
+
+  func callFlutterEchoNullable(_ aMap: [String? : Any?]?, completion: @escaping ([String? : Any?]?) -> Void) {
+    flutterAPI.echoNullable(aMap) { completion($0) }
+  }
+}
+
+enum ErrType: Error {
+  case thrownErrow
 }