+++ packages/pigeon/platform_tests/test_plugin/windows/test_plugin.h
@@ -82,9 +82,87 @@ class TestPlugin : public flutter::Plugin,
       std::function<
           void(std::optional<core_tests_pigeontest::FlutterError> reply)>
           result) override;
+  void CallFlutterEchoAllTypes(
+      const core_tests_pigeontest::AllTypes& everything,
+      std::function<
+          void(core_tests_pigeontest::ErrorOr<core_tests_pigeontest::AllTypes>
+                   reply)>
+          result) override;
+  void CallFlutterSendMultipleNullableTypes(
+      const bool* a_nullable_bool, const int64_t* a_nullable_int,
+      const std::string* a_nullable_string,
+      std::function<void(core_tests_pigeontest::ErrorOr<
+                         core_tests_pigeontest::AllNullableTypes>
+                             reply)>
+          result) override;
+  void CallFlutterEchoBool(
+      bool a_bool,
+      std::function<void(core_tests_pigeontest::ErrorOr<bool> reply)> result)
+      override;
+  void CallFlutterEchoInt(
+      int64_t an_int,
+      std::function<void(core_tests_pigeontest::ErrorOr<int64_t> reply)> result)
+      override;
+  void CallFlutterEchoDouble(
+      double a_double,
+      std::function<void(core_tests_pigeontest::ErrorOr<double> reply)> result)
+      override;
   void CallFlutterEchoString(
       const std::string& a_string,
       std::function<void(core_tests_pigeontest::ErrorOr<std::string> reply)>
+          result) override;
+  void CallFlutterEchoUint8List(
+      const std::vector<uint8_t>& a_list,
+      std::function<
+          void(core_tests_pigeontest::ErrorOr<std::vector<uint8_t>> reply)>
+          result) override;
+  void CallFlutterEchoList(
+      const flutter::EncodableList& a_list,
+      std::function<
+          void(core_tests_pigeontest::ErrorOr<flutter::EncodableList> reply)>
+          result) override;
+  void CallFlutterEchoMap(
+      const flutter::EncodableMap& a_map,
+      std::function<
+          void(core_tests_pigeontest::ErrorOr<flutter::EncodableMap> reply)>
+          result) override;
+  void CallFlutterEchoNullableBool(
+      const bool* a_bool,
+      std::function<
+          void(core_tests_pigeontest::ErrorOr<std::optional<bool>> reply)>
+          result) override;
+  void CallFlutterEchoNullableInt(
+      const int64_t* an_int,
+      std::function<
+          void(core_tests_pigeontest::ErrorOr<std::optional<int64_t>> reply)>
+          result) override;
+  void CallFlutterEchoNullableDouble(
+      const double* a_double,
+      std::function<
+          void(core_tests_pigeontest::ErrorOr<std::optional<double>> reply)>
+          result) override;
+  void CallFlutterEchoNullableString(
+      const std::string* a_string,
+      std::function<void(
+          core_tests_pigeontest::ErrorOr<std::optional<std::string>> reply)>
+          result) override;
+  void CallFlutterEchoNullableUint8List(
+      const std::vector<uint8_t>* a_list,
+      std::function<void(
+          core_tests_pigeontest::ErrorOr<std::optional<std::vector<uint8_t>>>
+              reply)>
+          result) override;
+  void CallFlutterEchoNullableList(
+      const flutter::EncodableList* a_list,
+      std::function<void(
+          core_tests_pigeontest::ErrorOr<std::optional<flutter::EncodableList>>
+              reply)>
+          result) override;
+  void CallFlutterEchoNullableMap(
+      const flutter::EncodableMap* a_map,
+      std::function<void(
+          core_tests_pigeontest::ErrorOr<std::optional<flutter::EncodableMap>>
+              reply)>
           result) override;
 
  private: